press z for rotate mirror
I made this mechanic so far;
The mirror cube is like a character now and it can not reflect laser.
The mirror cube can go to a reflector mirror, when they collide, if you press x then reflector starts to follow the mirror cube (if you press x again it stops following) and if you press z, reflector rotates 360 degree and then you try to manipulete laser.